Revert "use TextField CVR-647"
