+++ src/pages/policies/new.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-import { Breadcrumb } from 'components'
+import { Breadcrumb, Description } from 'components'
 import { MAX_INPUT_LENGTH as maxlength } from 'components/const'
 import { entityCodes, loadEntityCodes } from 'data/entityCodes'
 import { createPolicy } from 'data/policies'
@@ -7,7 +7,7 @@ import { formatPageTitle } from 'helpers/pageTitle'
 import { policyDetails } from 'helpers/routes'
 import { assertHas } from '../../validation/assertions'
 import { goto, metatags } from '@roxi/routify'
-import { Button, SearchableSelect, TextField, Page, setNotice, Form } from '@silintl/ui-components'
+import { Button, SearchableSelect, TextField, Page, setNotice } from '@silintl/ui-components'
 import { onMount } from 'svelte'
 
 let account = ''
@@ -61,48 +61,39 @@ const onCheck = (e: CustomEvent) => {
 }
 </script>
 
-<style>
-/* TODO use tailwind classes for this */
-.extra-margin {
-  margin-top: 1.5rem;
-  margin-bottom: 1.5rem;
-}
-</style>
-
 <Page>
   <Breadcrumb />
-  <Form on:submit={onCreatePolicy}>
-    <div>
-      <TextField label="Policy name" description="A name for the policy" {maxlength} required bind:value={policyName} />
-    </div>
 
-    <div class="extra-margin">
-      <SearchableSelect
-        placeholder="Entity"
-        options={entityOptions}
-        bind:choice={entityCodeName}
-        on:check={onCheck}
-        on:chosen={(e) => (entityCode = e.detail)}
-      />
-    </div>
+  <p>
+    <span class="header">Policy name<span class="required-input">*</span></span>
+    <TextField {maxlength} required autofocus bind:value={policyName} />
+  </p>
 
-    <div>
-      <TextField {maxlength} required description="e.g., ABCD12" label="Cost center" bind:value={costCenter} />
-    </div>
+  <p>
+    <span class="header">Entity<span class="required-input">*</span></span>
+    <SearchableSelect
+      options={entityOptions}
+      bind:choice={entityCodeName}
+      on:check={onCheck}
+      on:chosen={(e) => (entityCode = e.detail)}
+    />
+  </p>
 
-    <div>
-      <TextField {maxlength} required label="Account" description="e.g., 12345" bind:value={account} />
-    </div>
+  <p>
+    <span class="header">Cost center<span class="required-input">*</span></span>
+    <TextField {maxlength} required label="ABCD12" bind:value={costCenter} />
+  </p>
 
-    <div>
-      <TextField
-        label="Account Detail"
-        description="Appears in your statements"
-        {maxlength}
-        bind:value={accountDetail}
-      />
-    </div>
+  <p>
+    <span class="header">Account<span class="required-input">*</span></span>
+    <TextField {maxlength} required label="12345" bind:value={account} />
+  </p>
 
-    <Button raised>Create policy</Button>
-  </Form>
+  <p>
+    <span class="header">Account Detail</span>
+    <TextField label="details" {maxlength} bind:value={accountDetail} />
+    <Description>Appears in your statements</Description>
+  </p>
+
+  <Button raised on:click={onCreatePolicy}>Create policy</Button>
 </Page>
